+++ src/2020/buri/aoc/y20/d22/Day22Test.java
@@ -27,7 +27,7 @@ public class Day22Test extends BaseTest {
 	 */
 	@Test
 	public void testPart1Puzzle() {
-		int result = Day22.getResult(Part.ONE, Day22.getInput(0));
+		long result = Day22.getResult(Part.ONE, Day22.getInput(0));
 		toConsole(result);
 		assertEquals(0, result);
 	}
@@ -42,7 +42,7 @@ public class Day22Test extends BaseTest {
 	 */
 	@Test
 	public void testPart2Puzzle() {
-		int result = Day22.getResult(Part.TWO, Day22.getInput(0));
+		long result = Day22.getResult(Part.TWO, Day22.getInput(0));
 		toConsole(result);
 		assertEquals(0, result);
 	}
